+++ src/Symfony/Component/EventDispatcher/EventDispatcher.php
@@ -43,7 +43,6 @@ class EventDispatcher implements EventDispatcherInterface
             $event = new Event();
         }
 
-        $event->setDispatcher($this);
         $event->setName($eventName);
 
         if (!isset($this->listeners[$eventName])) {