+++ src/Symfony/Component/EventDispatcher/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+3.0.0
+-----
+
+  * The methods Event::setDispatcher(), Event::getDispatcher() have been removed.
+    The event dispatcher is passed to the listener call.
+
 2.5.0
 -----